[renderer] Clean up viewport setting
Software is still a mess, and vulkan never supported viewsize, but
otherwise everything seems fine.

+++ libs/video/renderer/vid_render_glsl.c
@@ -375,10 +375,10 @@ glsl_bind_framebuffer (framebuffer_t *framebuffer)
 static void
 glsl_set_viewport (const vrect_t *view)
 {
-	float x = view->x;
-	float y = vid.height - (view->y + view->height);
-	float w = view->width;
-	float h = view->height;
+	int         x = view->x;
+	int         y = vid.height - (view->y + view->height);
+	int         w = view->width;
+	int         h = view->height;
 	qfeglViewport (x, y, w, h);
 }
